pyproject.toml
[project]
name = "piper-tts-reader"
version = "0.1.0"
description = "Cross-platform menu bar application for reading text aloud using Piper TTS"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "piper-tts>=1.2.0",
    "pystray>=0.19.0",
    "Pillow>=9.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
    "pydub>=0.25.0",
    "pynput>=1.7.6",
    "structlog>=24.1.0",
    "svglib>=1.6.0",
    "reportlab>=4.4.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = []
